feat(module:tree): support virtual scroll close #4426 close #3808 close #3436 close #2680 close #1771
