+++ components/core/tree/nz-tree-base.definitions.ts
@@ -27,6 +27,5 @@ export interface NzFormatBeforeDropEvent {
 }
 
 export interface NzTreeNodeBaseComponent {
-  setClassMap(): void;
   markForCheck(): void;
 }
