+++ components/tree/demo/directory.md
@@ -7,8 +7,8 @@ title:
 
 ## zh-CN
 
-自定义目录树(使用 `nzTreeTemplate` 实现，支持右键)。
+使用 `nzTreeTemplate` 实现自定义目录结构，通过 `let-origin="origin"` 获得原始数据，`let-node` 获取当前节点状态。
 
 ## en-US
 
-Customize directory tree(with `nzTreeTemplate`, supported contextmenu).
+Customize directory tree with `nzTreeTemplate`, get data from `let-origin="origin"`, get node status from `let-node`.
