+++ components/tree/demo/virtual-scroll.md
@@ -1,0 +1,14 @@
+---
+order: 8
+title:
+  zh-CN: 虚拟滚动
+  en-US: Virtual Scroll
+---
+
+## zh-CN
+
+设定 `nzVirtualHeight` 开启虚拟滚动。
+
+## en-US
+
+Set `nzVirtualHeight` to enable virtual scroll.
